Composer command should specify php.ini location
Currently the Composer plugin does not include "-c <php.ini-location>"
parameter in the command that execute the composer.phar in an external
process.

This may cause troubles if multiple PHP installations are available on
the system. There may be installations (like Zend Server) which register
the php.ini file in the Windows Registry. This takes precedence over
php.ini in the same folder as the php.exe. Hence, if Composer is
configured to run with another PHP executable which has a php.ini in the
php.exe folder, it will actually be executed with the php.ini of Zend
Server. This causes lots of problems, like Composer miserably fails.

The solution is to include "-c <php.ini-location>" parameter in the
command. Which this patch actually does using the handy
PHPINIUtil.findPHPIni() method from PDT to locate the php.ini file.